refactor: Merge EditCampForm and EditCampPage into CampEditPage for simplified structure and improved maintainability
